marginal -> float

+++ UnBBayes/src/unbbayes/example/GraphicMode.java
@@ -103,9 +103,9 @@ public class GraphicMode {
                    System.exit(1);
                 }
 
-                double likelihood[] = new double[auxVP.getStatesSize()];
-                likelihood[0] = 1.0;
-                likelihood[1] = 0.8;
+                float likelihood[] = new float[auxVP.getStatesSize()];
+                likelihood[0] = 1f;
+                likelihood[1] = 0.8f;
 
                 auxVP.addLikeliHood(likelihood);
                 try {
